requirements.txt
pikepdf
reportlab
pillow
pytesseract
pdf2image
pdfplumber
python-docx
openpyxl
fpdf
PyMuPDF